pyproject.toml
[tool.poetry]
name = "bottoolkit"
version = "0.0.1"
description = ""
authors = ["Joao Cabral <jpfcabral@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
botbuilder-core = "^4.14.8"
aiohttp = "^3.9.3"
botbuilder-dialogs = "^4.14.8"
loguru = "^0.7.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
